refactor(HolyGrailLayout): move from features/auth/components to common/components
